requirements.txt
Flask
flask-jwt-extended
Flask-SQLAlchemy
psycopg2-binary
dataclasses-json
pytest